pyproject.toml
[tool.poetry]
name = "arg-gpt"
version = "0.0.1"
description = "A package to help you build a GPT interface for your python code"
authors = ["Philip Krejov"]
readme = "README.md"
packages = [{include = "arg_gpt"}]

[tool.poetry.dependencies]
python = "^3.8"
openai = "*"
typer = "*"
python-dotenv = "*"
groq = "*"

[tool.poetry.scripts]
arg-gpt = "examples.example_cmd:run_arg_prompt"
arg-gpt-groq = "examples.example_groq:run_arg_prompt"

[tool.poetry.group.dev.dependencies]
pytest-cov = "5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
